Update: One new function & Dicts
1. Added one new function: Count_MaxorMin. This new function takes has two parameters. The first is an integer which identifies which operation to perform. Only a 1 or 2 is accepted. 1 = min function. 2 = max function. The second parameter is an array of number values. The array must be one-dimensional and can vary in length. The output of the function is an integer which depicts how many times the "minimum" or "maximum" value in the array is counted, e.g., it informs the user if there are duplicate "minimum" or "maximum" values.

E.g., if the minimum value of an array = 4 and this same array contains three 4 values, then this function will return 3. This returned value informs the user that the minimum value of their array (which is 4), exists in their array 3 times, i.e., the minimum value is duplicated.

2. Made progress to fuzzy_match() by writing to do the following:

First: Identify if the string being evaluated as a possible match is a phrase or sentence (basically, if its more than one word)

Second: If the "string" is actually more than one word, a conditional sequence will then split that sentence into substrings (1 substring = 1 word), and then evaluate the levenshtein distance (edit distance) and string similarity of each substring against the target string (targetS - the string we are seeking a match for)

Moving forward: still need to finish by resolving conflicts when strings share the same minimum edit distance and tie in the results of the substring evaluation with the overall fuzzy_match() function.